rename old server.js to server_old.js, new server.js uses express.js, ignore umass_reserve_rooms folder (local db)
